Fix padding for body and avoid shrinking
When the browser width would get too small, the header elements would
begin to overflow to the next line, so to avoid this, I added
flex-shrink: 0 to the header elements as well as the sign-up-banner
button.

Additionally, I tweaked the margins between the links in the header for
aesthetic reasons.